Started refactoring user + oauth functionality into blueprints
